This is my Script.js file. In here i made sure to include latitude and longitude for our Weather API.
I used the following website: 
open-meteo.com to get the following, 
latitude and longitude 

I also include a fetch, so that the current weather can be fetched. 

And last I Included a section to display the weather data. So kind of like a form so you can see what output you get from the API